pyproject.toml
[tool.poetry]
name = "fastapi-skeleton"
version = "0.1.0"
description = "Providing a basic fastapi application to start your projects with a simple skeleton."
authors = [
  "Dominik Geldmacher <oryon@cyberise.de>",
]
readme = '../README.md'


[tool.poetry.dependencies]
python = "^3.10.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.8.2"}
fastapi = "^0.68.0"
httpx = "^0.19.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
devtools = "^0.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
humanize = "^3.11.0"
databases = {extras = ["sqlite"], version = "^0.5.2"}
alembic = "^1.7.1"
sqlmodel = "^0.0.4"
pytest-cov = "^2.12.1"
coverage = "^5.5"
gunicorn = "^20.1.0"
rich = "^10.11.0"
dill = "^0.3.4"
pymemcache = "^3.5.0"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
datamodel-code-generator = "^0.11.12"
black = "^21.8b0"
ipykernel = "^6.3.1"
pytest-env = "^0.6.2"
flake8 = "^3.9.2"
typer = "^0.4.0"
mkdocs = "^1.2.2"
mkdocs-material = "^7.3.0"
mkdocs-render-swagger-plugin = "^0.0.3"
pytest-asyncio = "^0.15.1"
pytest-mock = "^3.6.1"


[tool.black]
line-length = 119
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | build
    | dist
    | htmlcov
    | migrations
  )/
)
'''


[tool.flake8]
statistics = "True"
ignore = ["D203" , "W503"]
exclude = '''
    .git,
    '''
max-line-length = 119


[tool.pytest.ini_options]
minversion = "6.2.5"
addopts = "-p no:warnings"
python_files = "tests.py test_*.py *_tests.py"
norecursedirs = ".envs .git compose htmlcov migrations notebooks resources"
env = [
    "TESTING = True",
]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
